generate requirements.txt from setup.py using pip-compile
part of https://github.com/jazzband/pip-tools

It's better to separate the actual requirements (from setup.py) from the other test- or dev-related requirements.
We store the latter in a separate dev-requirements.in file, and will generate a dev-requirements.txt from it using pip-compile.

pip-compile is basically a better version of pip freeze.