remove pinned requirements and always test with latest transitive deps
The setup was getting overly complicated. It's OK to test the latest which is what the user will get when they pip install us. That way we get notified immediately if anything breaks. And it reduces the information overload for future contributors.

Same as https://github.com/googlefonts/picosvg/pull/161