only use proportional advance width when larger than default width
Previously, I was simply scaling the FontConfig.width by the aspect ratio. But that can produce unintended consequences if the default width != font height (ascender-descender): e.g. if one wanted to pad glyphs on both sides (like we do with noto-emoji) by setting a wider default width than the font height, the impact of non-square viewbox on that padding would be even greater as the viewbox gets wider.
It's better to do this: given the font height and the viewbox aspect ratio, scale font heigth by aspect ratio to obtain the scaled advance width; ise the latter if it's larger than the default width, otherwise use the default width (and place things centered).
A user can play with the default --width (even set it to 0) until they like it. This effectively becomes the minimum default advance width, to be used only used when the proportional advance width that is computed from the viewbox width (relative to its height which is fixed to ascender-descender) is smaller. Yeah, I think I like that.