when reusing inter-glyph shapes, reuse any intra-glyph shapes too
For --color_format=picosvg we try to group glyphs that use same shapes in a single svg document so we can reuse shapes across multiple glyphs. We also reuse paths within a given glyph (e.g. two ellipses for the eyes of a smiley emoji).
Now there was actually a bug in the current code: we were only applying the intra-glyph sharing for the first occurrence of a given painted layer in a color glyph, but not again when the same painted layer is reused by a different color glyph.

Our clocks test shows what I mean by this. It has two input svg, one-o-clock and two-o-clock. The one-o-clock has the same grey dot repeated four times (an example of intra-glyph reuse). The two-o-clock similary has the same four dots (same painted layer with internal reuse repeated in a different color glyph). Before this patch, the two-o-clock would only contain one grey dot and would miss the rest of them that are repeated.