nanoemoji/svg.py: fix issue with id-less elements in defs
Fixes #346

Make sure we don't append to <defs> a reused_el which is not a <path> but actually a <use> element (that point to that path).
This prevents ending up with id-less <use> elements in <defs>, which
(via an complicated code path) lead to a KeyError when attempting to
sort <defs> by 'id' attribute.