WIP: fix issue with id-less elements in defs
when I try to build noto-emoji --color_format=picosvg, I get a hard crash when nanoemoji tries to sort the <defs> elements by id. I have a feeling I'm on to something but I still haven't managed to isolate a reproducer.
The issue seems to be related to _migrate_to_defs replacing a parent-less reused_el with its <use> in reuse_cache.glyph_elements (I thought #346 could not happen but it _does_ when building noto-emoji, not sure how), and then running _migrate_to_defs again with the reused_el this time being a <use> and no longer a <path>, so we end up with <use> elements in <defs>.. honestly the code is quite complicated.
With this patch, at least it doesn't crash any more while trying to sort <defs> by id. But I still have to fully understand what's going on.