workaround for Skia bug whereby linear gradient's P2 has no effect
because of a regression in Skia, skewed linear gradients (i.e. where P2 is not perpendicular to P1-P0 vector) don't render correctly in Chrome.
The position of rotation point P2 has no effect and the gradient never gets rotated.
As workaround, we normalize the linear gradients so that they will look the same whether or not P2 is taken into account, by setting P1 to its projection onto the line perpendicular to the rotation vector.
We can remove this once the bug is fixed in Chrome.

https://skia-review.googlesource.com/c/skia/+/494676
https://bugs.chromium.org/p/chromium/issues/detail?id=1287162