use abspath instead of Path.resolve() so symlinks aren't resolved
noto-emoji uses symlinks for a couple region flags that are indentical. Currently our use of Path.resolve() to make paths absolute leads to resolving the symlinks and then hard crash because of duplicate glyphs. If we just make the paths absolute without resolving the symbolic links we can support this use case without other changes.