Use .source_names.txt file to generate glyphmap
this was left unused for a while. Originally it was used by write_codepoints rule (later replaced by write_glypmap).
I think it suits perfectly the use case of generating the GlyphMappings, because it's only the stem of the source filenames, not their directory nor their suffix, that should inpact on the respective codepoints and glyph names.
.glyphmap files can't vary between masters, because masters have the same source names and glyph names, and these must map to the same codepoints. They can only vary between different font configs, if these specify different glyphmap generator scripts.